add KerasObject, KerasLayer subclass it
