+++ syft_tensorflow/tensor/__init__.py
@@ -1,4 +1,6 @@
 from syft_tensorflow.tensor.tensor import TensorFlowTensor
 from syft_tensorflow.tensor.variable import TensorFlowVariable
+from syft_tensorflow.tensor.keras_object import KerasObject
 from syft_tensorflow.tensor.keras_layer import KerasLayer
 from syft_tensorflow.tensor.keras_model import KerasModel
+
